+++ src/context/AuthContext.tsx
@@ -1,47 +1,5 @@
-import type { ReactNode } from 'react';
-import React, { createContext, useContext, useEffect, useState } from 'react';
+import { createContext } from 'react';
 
-interface AuthContextProps {
-  authToken: string | null;
-  login: (token: string) => void;
-  logout: () => void;
-}
+import type { AuthContextProps } from './types';
 
-interface AuthProviderProps {
-  children: ReactNode;
-}
-
-const AuthContext = createContext<AuthContextProps | undefined>(undefined);
-
-export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
-  const [authToken, setAuthToken] = useState<string | null>(sessionStorage.getItem('authToken'));
-
-  const login = (token: string) => {
-    setAuthToken(token);
-    sessionStorage.setItem('authToken', token);
-  };
-
-  const logout = () => {
-    setAuthToken(null);
-    sessionStorage.removeItem('authToken');
-  };
-
-  useEffect(() => {
-    const token = sessionStorage.getItem('authToken');
-    if (token) {
-      setAuthToken(token);
-    }
-  }, []);
-
-  return (
-    <AuthContext.Provider value={{ authToken, login, logout }}>{children}</AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
-};
+export const AuthContext = createContext<AuthContextProps | undefined>(undefined);
